Added percentage to group, made it two decimals

+++ code/infra/frontend-adapters/rest-adapter/src/main/java/be/tr/democracy/rest/PartyVotesViewDTO.java
@@ -1,5 +1,5 @@
 package be.tr.democracy.rest;
 
-public record PartyVotesViewDTO(String partyName, int votePercentage, int numberOfVotes) {
+public record PartyVotesViewDTO(String partyName, double votePercentage, int numberOfVotes) {
 
 }
